+++ flake.nix
@@ -123,7 +123,7 @@
           test-playwright = import ./nix/devenv-modules/tasks/shared/test-playwright.nix;
           storybook = import ./nix/devenv-modules/tasks/shared/storybook.nix;
           netlify = import ./nix/devenv-modules/tasks/shared/netlify.nix;
-          lint-effect-lsp = import ./nix/devenv-modules/tasks/shared/lint-effect-lsp.nix;
+          ts-effect-lsp = import ./nix/devenv-modules/tasks/shared/ts-effect-lsp.nix;
           vercel = import ./nix/devenv-modules/tasks/shared/vercel.nix;
           lint-oxc = import ./nix/devenv-modules/tasks/shared/lint-oxc.nix;
           bun = import ./nix/devenv-modules/tasks/shared/bun.nix;